[ASN-47] fix: 카테고리 선택 아코디언 컴포넌트 분리
